added pubnub event updates
